:shirt: Refactor formats to move converter code
