+++ src/Ekona/Containers/Converters/Fat.cs
